NCBC-3676: Force metadata to be read before iteration query rows
Motivation
----------
This commit forces the enumeration to read ahead before the app
enumerates the result. The person is to ensure that the metadata
is read and if a RpcException will be thrown, it will be thrown
in the RetryHandler instead of when the results are enumerated
by the app. This is a foundation commit as there will be amendments,
for example retry considerations for various query responses.

Modifications
-------------
 - Force enumeration to read ahead withen the RetryHandler. Cache
the results and then enumerate them first if successful.
 - Throw a CouchbaseException if an unhandled status is returned

Change-Id: Idc98127bad7d7b922e0269609fe717aaa3f7b867
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/206064
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>