NCBC-3684: Add TryXXX methods for CRUD
Motivation
----------
The SDK throws DocumentNotFoundExceptions when a the server returns a
KeyNotFound status. There a performance benefit if the exception is not
thrown as there is overhead associated with generating and throwing the
exception. This commit add TryXxx methods for KV CRUD operations so that
the state of existance of a document can be checked instead of throwing
the exception.

Modifications
-------------
 - Added TryXxxAsync extension methods to CollectionExtensions
 - Added ITryXxxResult types and their implementations
 - Added basic functional tests

Change-Id: I6c04504fb34790103e3e18fa69fab30fbf856300
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/206305
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>