NCBC-3739: Send GetConfigMap operations without queuing
Motivation
----------
A GetConfigMap begin sent is generally a high priority because it
indicates server configuration is outdated due to failover or rebalance.
It should take precedence over general K/V operations or they will tend
to fail with more NMVB errors.

Modifications
-------------
Add a method, which currently only applies to the default
`ChannelConnectionPool`, which allows an operation to be jumped to the
front of the line to be sent. It randomly selects the connection from
the pool to provide distribution rather than relying on the queue to
distribute.

Use this method for all GetClusterMap operations.

Refactor GetRandom on a list to RandomOrDefault and reduce heap allocs
when the list size is unknown.

Results
-------
Under load when an NVMB or similar triggers a GetClusterMap operation
the operation will be prioritized and be sent as soon as reasonably
possible. However, the legacy DataFlowConnectionPool does not implement
it and will see no change in behavior.

Change-Id: I17194fd18af4b18385211394282bb547eea99946
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/207582
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>