NCBC-3742: Reduce push config spam in ConfigPushHandler
Motivation
----------
ConfigVersions are sent to the ConfigPushHandler and then
the config is fetched and sent to the ConfigHandler. The
ConfigHandler is getting too many duplicate which is causing
a backup meaning buckets are getting far behind the config rev.

Modifications
-------------
 - Fix bug where equal config version are not skipped
 - After fetching a candidate config, check to see if
its revision is before the current revision before sending
it to the ConfigHandler.

Change-Id: Icbb95dc3a7bd0436724dde40b7f506efbd3ae505
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/207673
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>