NCBC-3747: Apply network config to cluster map when pushed to SDK
Motivation
----------
When CB Server 7.2 sends a cluster map changed notification, it is sent
with a JSON body containing the new map. After deserializing, we are not
replacing $HOST with the correct values, causing the in-memory map to
become corrupt and subsequent operations to timeout.

Modifications
-------------
After deserializing a pushed notification with a cluster map replace
$HOST with the endpoint and apply any overridden network resolution
settings. This matches the behavior when we receive a BucketConfig back
from an explicit request.

Results
-------
Couchbase Server versions which push config map bodies but which don't
supply extras with the config version no longer corrupt the SDKs
internal cluster map.

Change-Id: Ifa38563160d8cf9f528e746c5e101abbfdbc839e
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/207789
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>