NCBC-3758: PersistentQueue does not dequeue items in original FIFO order
Motivation
----------
This bug was introduced in SDK 3.4.15 following this change NCBC-3578.

A bug in the PersistentQueue datastructure dequeues items in LIFO order (removes item at index -1), and returns the first item (at index 0) on each Dequeue().

Changes
-------
Dequeue now targets the first element of the queue rather than the last.

Change-Id: I2ca7a7eb7e7b2626513c82b7ade80df602321479
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/208536
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>