NCBC-3792: Tune in-flight operation limits to perform better at high latency
Motivation
----------
High latency connections may require a higher in-flight operation limit
in order to acheive throughput. However, the current backing
implementation of InFlightOperationSet is hyper-optimized for a very low
limit and performs poorly at high limits. Currently, both adding and
in-flight operations and looking up in-flight operations is a O(n)
operation.

Modifications
-------------
- Back InFlightOperationSet with a ConcurrentDictionary.
- Increase the default limit from 8 to 16 operations.

Results
-------
This is probably a very slight performance hit in the small case due to
lock contention and hash table overhead. However, the hash table won't
fall off a performance cliff with large numbers of items.

Change-Id: I3d4c660b3fa58d210c5ba0a4ee674ebc727447ca
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/210464
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Emilien Bevierre <emilien.bevierre@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>