NCBC-3710: Tag metrics with outcome
Motivation
----------
The Metrics RFC states that operation latency metrics should include an
"outcome" tag indicating the success or failure of the operation. The
Java SDK outputs either "Success" or an error string. The .NET SDK
doesn't currently output this tag.

Modifications
-------------
- Add an `errorType` parameter to the metric helpers for duration
  metrics and use this to generate an outcome tag in a manner consistent
  with the Java SDK.
- Forward this from new StopRecording overloads on `IRequest`, using
  default implementations to be as backward-compatible with custom
  `IRequest` or `RequestBase` implementations as possible.
- Fill this `errorType` on any error scenario (exception or handled
  KeyNotFound status) in the RetryOrchestrator, half-open circuit
  breaker tests, and various internal ClusterNode operations.

Results
-------
All operations now report an outcome of "Success" or an error code as a
tag on their duration metrics, in alignment with the RFC and the Java
SDK.

Change-Id: I76b425a1442f07048a69f3398667340dee00189a
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/212396
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Emilien Bevierre <emilien.bevierre@couchbase.com>