NCBC-3780: FIT TXN: SingleQueryTransactions failures (26)
Motivation
----------
Tons of tests in the SingleQueryTransaction suite were failing due to the SDK throwing the wrong exception when executing nested SingleQueryTransactions with a FailHard hook.

The source of the issue was that QueryClient was cloning the QueryOptions when retrying a query, but wasn't cloning QueryContext. This brings up the other issue that RetryOrchestrator shouldn't be retrying transaction requests with the same handling as normal SDK operations.
It would be worth instantly getting a ReadOnly record of the QueryOptions once the user hands it to the SDK to make the data immutable and prevent any potential threading issues, but that record isn't "complete" due to an inconsistent accessiblity issue.

Change-Id: I5384291c7ca5cc08be151327c562bad5ff8c44bd
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/211075
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>