TXNN-131: Transactions - support ExtInsertExisting
Motivation
----------
This simple extension ensures ctx.insert() consistently raises a DocumentAlreadyExistsException, which the user can choose to ignore and have the transaction continue. See CBD-4140 for details.

Changes
-------
In AttemptContext.CreateStagedInsert, if DocumentExists is caught, we now let it bubble up the chain instead of wrapping it in a TransactionOperationFailedException.

Change-Id: I4b015262651c92509c0a360ba3bcd7e580f7c430
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/211194
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>