NCBC-3837: Add option to disable duplicative legacy metrics
Motivation
----------
Currently enabling OpenTelemetry metrics produces some metrics that are
duplicative of the primary metrics now that proper tagging is in place.
This can create additional costs for metric infrastructure or SaaS
providers. We should provide an easy means to simplify the output
metrics, but maintain backward compatibility for current consumers that
may rely on these metrics.

Modifications
-------------
- Add an overload that accepts an action to configure
  CouchbaseMeterInstrumentationOptions
- If CouchbaseMeterInstrumentationOptions is provided and
  ExcludeLegacyMetrics is true, drop the duplicative metrics
- Enable nullable reference types in the OTel package
- Improve integration tests and allow for testing to .NET Aspire
  dashboards
- Remove custom Snappier wireup and use the newer built-in option
- Move NonParallelDefinition to Couchbase.Test.Common
- Add a unit test and fixup an existing unit test

Results
-------
No change unless the consumer explicitly requests that legacy metrics be
excluded, in which case these metrics are dropped and not sent to the
configured OTel exporter.

Change-Id: I26b365e5ab54889365e8f01c58299e25b126b3ae
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/213438
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Emilien Bevierre <emilien.bevierre@couchbase.com>
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>