NCBC-3845: ReadOnlySequence and IBufferWriter support in ITypeSerializer
Motivation
----------
.NET 9 includes a new HybridCache model which blends an L1 memory cache
with an L2 distributed cache. This includes back-compat to .NET older
runtimes. Couchbase is an ideal solution for the L2 distributed cache.

Included in the hybrid cache is the IHybridCacheSerializer<T>. To
plug-in the Couchbase SDK ITypeSerializer as the hybrid cache
serializer, ITypeSerializer will need support for both deserializing
from a ReadOnlySequence<byte> and serializing to an IBufferWriter<byte>.
These are logically similar to our current support for
ReadOnlyMemory<byte> and Stream, but offer more efficiencies around
buffer management.

Modifications
-------------
- Add Utf8Helpers.TrimBomIfPresent overload that accepts a
  ReadOnlySequence<byte> and add unit tests for all overloads
- Refactor Utf8MemoryReader to work on ReadOnlySequence<byte> rather
  than ReadOnlyMemory<byte>
- Create BufferWriterStream to wrap an IBufferWriter<byte> in a Stream
  for Newtonsoft
- Add a new IBufferedTypeSerializer interface which deserializes from
  ReadOnlySequence<byte> and serializes to IBufferWriter<byte>
- IBufferedTypeSerializer also includes CanSerialize to test if the
  serializer is capable of handling a specific type
- Implement IBufferedTypeSerializer on DefaultSerializer and
  SystemTextJsonSerializer

Results
-------
The new Serialize and Deserialize overloads are available for all
frameworks by casting to the IBufferedTypeSerializer interface. This
will allow any serializer that implements IBufferedTypeSerializer to be
easily wrapped in an IHybridCacheSerializer<T> for caching.

The CanSerialize method is used by IHybridCacheSerializerFactory to
allow fallback to other registered serializers when using a
JsonSerializerContext-based serializer with limited type support.

There may be a minor performance impact to DefaultSerializer working
through the additional abstraction of ReadOnlySequence<T> when reading
from a continguous block of memory. However, it should be neglible and a
fraction of the gains we added recently when we dropped the use of
StreamReader in NCBC-3421.

Change-Id: I927d951d04659dc44824174a9c8cb7ff259ded37
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/219023
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>