NCBC-3846: Seal internal classes for devirtualization perf gains
Motivation
----------
When a class is sealed, the JIT compiler has the opportunity to
devirtualize calls to virtual members when it knows that the instance
must be of the sealed class. This reduces the cost of the method call
and may also allow inlining to further improve performance.

This gain is even greater in .NET 8 with the advent of Dynamic PGO,
which can perform guarded devirtualization and subsequent inlining in
cases where the JIT compiler can't be certain of the type but observes
that the type is often a single class during runtime profiling. It can
even devirtualize across interface dispatch.

Modifications
-------------
- Seal a wide variety of internal classes which don't have any
  subclasses, especially those used in hot paths or which have virtual
  members or inherit from types with virtual members
- Refactor unit tests to allow sealing DataFlowConnectionPool and
  ServiceUriProvider
- Drop unused classes SynchronizationContextExclusion and BucketContext

Change-Id: I683214564283a33f04b4f03d25675e4e675131b6
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/214947
Reviewed-by: Emilien Bevierre <emilien.bevierre@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>