NCBC-3849: Track orphaned metrics regardless of tracing
Motivation
----------
The current code for tracking orphaned operations checks to see if the
span CanWrite before tracking. However, tracking to metrics is also
affected by this. The result is that if distributed tracing is disabled
(due to overall tracing being disabled or the specific request not being
sampled) the metric is never incremented.

Modifications
-------------
Track to metrics even if the span is not writable.

Results
-------
Orphaned metrics are now tracked in all cases where the operation is
sent over the wire.

Change-Id: Id4d136b023ad87261d54f08258e2840c909fefba
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/215126
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Emilien Bevierre <emilien.bevierre@couchbase.com>