NCBC-3850: Track timeout metrics in ClusterNode.SyncAsync
Motivation
----------
Timeouts may occur during the operation, which are handled in
ClusterNode.SendAsync, or during the wait between retries, which are
handled by RetryOrchestrator.RetryAsync. The former is not posting to
the timeouts metric, only the latter.

Modifications
-------------
Post to the timeout metrics when timeouts are handled by ClusterNode.

Results
-------
Both timeout cases now correctly post to metrics.

Change-Id: I3cff2ec414b7d387f63e6e5dd4c1a9cb07cd36ce
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/215127
Reviewed-by: Emilien Bevierre <emilien.bevierre@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>