NCBC-XXXX: Dispose CTS when returned to pool on legacy frameworks
Motivation
----------
The pre-net6.0 implementation of the CancellationTokenSourcePool, which
simply mocks the pool since a CTS is not reusable on old runtimes,
is failing to Dispose of the CTS when it is returned. This can have
very sigificant performance impacts on old runtimes such as .NET 4.

Modifications
-------------
- Dispose of the CTS when returned
- Also don't capture the SynchronizationContext on operation
  subscriptions to cancellation tokens

Results
-------
Only legacy runtimes such as .NET 4 are affected.

- A very signficant performance improvement because timers related to
  the CTS are removed from the timer queue immediately rather than
  waiting for garbage collection. This impact is greater the higher the
  rate of operations.
- Removes a possible location for a deadlock in legacy ASP.NET in the
  case where a K/V operation times out but the request thread is
  blocked waiting on operation completion in sync-over-async usages.

Note: This benchmark is somewhat artificial because the rate of timeout
CTS cancellation relative to the rate of GC finalization is a factor.
The benchmark is renting/returning timeout CTSs far faster than a real
world application, and probably running GC at a different rate. This
means the runtime for the WithoutDispose test is probably inflated. But
it does still demonstrate the impact in a very high load scenario.

BenchmarkDotNet v0.13.12, Windows 11 (10.0.22631.4037/23H2/2023Update/SunValley3)
12th Gen Intel Core i7-1270P, 1 CPU, 16 logical and 12 physical cores
.NET SDK 8.0.400
  [Host]     : .NET 8.0.8 (8.0.824.36612), X64 RyuJIT AVX2
  Job-RIFZTT : .NET Framework 4.8.1 (4.8.9261.0), X64 RyuJIT VectorSize=256

Runtime=.NET Framework 4.8  Toolchain=net48

| Method         | Mean       | Error     | StdDev    | Ratio | RatioSD | Rank | Gen0   | Gen1   | Gen2   | Allocated | Alloc Ratio |
|--------------- |-----------:|----------:|----------:|------:|--------:|-----:|-------:|-------:|-------:|----------:|------------:|
| WithoutDispose | 1,649.9 ns | 162.30 ns | 478.53 ns |  1.00 |    0.00 |    2 | 0.0420 | 0.0143 | 0.0010 |     259 B |        1.00 |
| WithDispose    |   147.8 ns |   1.83 ns |   1.62 ns |  0.10 |    0.05 |    1 | 0.0343 |      - |      - |     217 B |        0.84 |

Change-Id: Ieaa31262aed3df5af17ec8c200db8a2e18e2b5fb
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/215751
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>