NCBC-3885 Use both provided certs and System truststore or just provided certs
Motivation
==========
Changes - make two attempts at validation. The first attempt uses the
system trust store, with the addition of the provided certs. The
provided certs maybe either ClusterOptions provided certs, or the
certs built-in to the SDK. If validation relies on the provided
certs, and the provided certs do not result in a valid chain, then
validation will fail with PartialChain or UntrustedRoot. If that
happens, a second attempt is made with the provided certs in a
CustomTrustStore. Since the CustomTrustStore comprises only certs
that were provided by the ClusterOptions, or certs built-into the
SDK - even if they are self-signed, they are trusted.

Modification
============
handle all the possibilities for certificates - system and
user-supplied, or Capella certificate. First try the system
truststore and user-supplied certificates. If that fails with
Untrusted or PartialChain, it is likely a Capella self-signed, so
try the Capella certificate by itself.

Results
=======
Any of the above possibilities for certificates
will succeed.

Change-Id: Ia72be580681b797bc32c93faf56ee54a82098e47
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/222027
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Emilien Bevierre <emilien.bevierre@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>