NCBC-3927 - Transactional GET logical change
Motivation
==========
Correct faulty logic when getting a doc that was in an abandoned transaction.

Modification
============
If the ATR doesn't exist, but there are staged changes, after you re-check, then
just return the original document.   The transaction that seems to be a blocking
txn never committed, so ignore the staged changes.

Result
======
Along with fixing the performer, you will see several tests pass in GettingDocumentsTest
in the FIT tests, that failed otherwise.

Change-Id: Id4b1c22d1fac5b840ebf8ed13dbabe352e20fb83
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/222163
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>