Changes for FIT to work
Needed to use interfaces for the hook classes, so as to minimize the
changes there.   Moved the constants and default functions into the DefaultTestHooks class, and
referenced them in the code that needs to.   Small change, but allows the
FIT to not have any major change.

Later I'd like to revisit this and try for a cleaner approach, but this
does work, and these are all internal so seems fine for now.

Change-Id: Ie2db0a807aa64e07cd315354bdff12d162aefe7d
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/222460
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: David Kelly <davidmichaelkelly@gmail.com>