NCBC-3948: Add new TryGetRawParameter to ClusterOptions
Motivation
----------
For EF Core we need to add the Bucket name and Scope to the connection string so that we have a unique connection string per DbContext. Since the CollectionString class which the ClusterOptions uses does not have a way of retrieving “unknown” parameters, we can write a method that allows for a raw parameter to be retrieved from the connection string.

Modifications
-------------
- Add TryGetRawParameter to ClusterOptions
- Add unit tests

Result
------
A raw unsupported by the SDK connection string value can be retrieved. This is useful in cases where an application using the SDK needs to append additional data to the connection string.

Change-Id: I9d528898fe1bc41c93736dec8fab230712f03f3b
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/223108
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>