NCBC-3937 Implement extSDKIntegration
Motivation
==========
Integrate the Transactions into the SDK.

Modification
===========
Cleanup now will not open all buckets and attempt to clean the
default collection and scope.   We now only try specific collections
which you specify in the TransactionsCleanupConfig or perhaps in
the TransactionsConfig (as just the metadataCollection).  In addition
we will add any collections where we have created metadata, which in
absence of a metadataCollection will be the default collection and
scope of the bucket associated with the first document which is
modified (insert, remove, replace).

Results
=======
Of the 3175 FIT tests, there are now 40 errors/failures.  A couple
are related to RBAC and cleanup, others are unrelated it seems...

Change-Id: I711fb43f4c12af23fa7af408a01c7db9f2374615
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/223245
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>