NCBC-3958 Cleanup Issues Need Addressing
Motivation
==========
We implemented sdkIntegration, and didn't quite fix all the cleanup issues.
Lets get the rest, as much as we can right now, before moving to
threadsafety.

Modification
============
We missed in sdkIntegration that we _always_ create a Transactions
instance when we create a Cluster.   The reason is the LostTxnCleanup
needs to be running as soon as it can (after bootstrap) since we
could have configured it to clean some collections when we configured
the Cluster.

Also when there was a lot of contention to create a new
PerCollectionCleaner, we were creating several for the same
collection.  ConcurrentDictionary will create values even when
it doesn't actually put the value in because someone else won the
race.  So use Lazy<T> instead, then that initializaiton will be
threadsafe.  <sigh>

Results
=======
Only one test failure in 102 cleanup tests.   Progress.

Change-Id: If2c08af7cd2f6c8013df0789b77f50f08ff33ffb
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/223515
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>