TXNN-130: extThreadSafe Implementation
Motivation
==========
To safely perform transaction ops in parallel with eachother, we
need to deal with some places where we have contention.  This will
allow us to do things like perform multiple inserts, get/replace,
etc.. in parallel with eachother for speedier transactions.

Modifications
=============
The transaction documentation has the overall approach.   In brief,
we need to lock the creation of the atr when we mutate the first
document.   Then we need to carefully switch over to queryMode,
making sure all the kv ops are complete and no new ones start
until we successfully switch over (at which point they will be
executed via Query).  Also queries actually are run in serial,
the server has a small buffer so one at a time is the only way.
The other place where we will have locking is around the commit
and rollback - both need to wait until there are no in-flight
operations.  Along the way, there are error handling changes
as well.

Results
=======
All the threadsafe tests in fit now pass.

Change-Id: I18725c3a3df470685ffbe310c8b546998912f2b8
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/224432
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>