NCBC-3902: Added a TryLookupInAsync method to CollectionExtensions
Motivation
==========
This is useful in the same sense as the TryGetAsync and others - we
gain some efficiency when not raising an exception and returning an
object with DocExists set to false instead.

Modifications
=============
Since Exists(int) is a method on LookupInResult, there is a
DocumentExists method on all the Try<Op>Result objects, and Exists
currently will work as well, except for TryLookupInResult, where
the Exists(int) hides it.   In the future we can deprecate and
remove Exists. For now all the Try<Op>Results have both Exists
and DocumentExists,

Results
=======
Works fine when testing locally.  Could consider some integration
tests for these in the future perhaps.

Change-Id: I2989dfefcf3c17c0192921dbd950af91329a367c
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/224567
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>