NCBC-3932 OOM during streaming query results
Motivation
==========
The point of streaming them is to not OOM, so lets get this looked
at.

Modification
============
When iterating with GetIteratorAsync, we need to get the previous
rows to GC sooner.   Seems that an async foreach() does this but
when using the iterator, it doesn't.   So... we set the row to
default(T) after yielding it, and that appears to do it.

Also - our query metrics had a uint for the ResultSize.   That
blows up at 4M, and we support 20Mb docs.   Needed to switch to
a ulong (64-bit unsigned int).

Results
=======
No more OOM

Change-Id: I656d8e859d3a600e07acad15d95de52b3f7b8bf6
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/224640
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>