TXNN-131 Support ExtInsertExisting
Motivation
==========
Catch up to other transaction implementations.   This one, we make
sure we throw a DocumentAlreadyExists exception when trying to
insert a document which is already present.  If you don't catch
it, then the transaction will rollback, but you can catch and
ignore.  This seems niche, but some use case exists for it.

Modification
============
We always raise it when attempting to insert a document that is
already present, or one which is being inserted in another transaction.

Results
=======
ExtInsertExisting tests pass, along with others that check for the
DocumentAlreadyExists exception when the performer declares it
supports ExtInsertExisting.

Change-Id: I13ad8e33665908d41ef57dc4d3a208c57ccf9122
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/225049
Reviewed-by: Emilien Bevierre <emilien.bevierre@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>