NCBC-3985: Add ReviveDocument to MutateInOptions
Motivation
==========
Added in 7.1, this is a server feature.   It is important for the
extReplaceBodyWithXattr extension to transactions, which is a
efficiency improvement.   So lets do it.

Modification
============
We added it as a bucket capability, and added it as another option
in the MutateInOption class.  Then, in MutateInAsync, we check for
it in the BucketCapabilities when using it.   Also - this expects
the AccessDeleted flag, so we add it when we add this one.  Also
a new exception if you try to revive a document when it is not
a tombstone.

Results
=======
Added tests, they pass.
in MutateInAsync.

Change-Id: I2e1158936aafd6f01396eec0fc56782f61a924e5
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/225219
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>