NCBC-3974: Adding profile to TransactionQueryOptions
Motivation
==========
A FIT test expects the Profile.Off to be set (even though it
is the default), when we set it in the options.   However, we
consider that the default and didn't send it.  This changes that
to always send the profile field.  One can argue that sending
this makes less assumption about the server, and follows a
principal of sending what the client wants to be sent.

Modification
============
Don't exclude the default value when serializing the query options

Result
======
Allows the FIT to pass without modification.

Change-Id: I0ce7cbbd0c2b0be249c43e896d727c2d26f39a68
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/225533
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>