NCBC-3957: Implement extReplaceBodyWithXattr
Motivation
==========
This is an optimization for clusters that support the ReviveDocument
and ReplaceBodyWithXattr flags.   We don't have to store the
document contents in our staged mutations in memory and move it
across to the server when unstaging.  Now we just tell the server
to move the already staged contents across to the body of the doc.

Modification
============
We just check to see if the server supports the flags, and use them.
There are some error handling changes that are part of this as
well - most of the changes have to do with that.   See the docs
on transactions for details.

Results
=======
All the extReplaceBodyWithXattr tests in FIT now pass.

Change-Id: I49da87fbc4a4a693d246919c3a06f4291538983c
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/225606
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>