NCBC-3993: Fail on certificate name failure when not ignored.
Motivation
==========
Validation was passing when certificate name validation failed.
If the built-in certificate validation failed, validation is
then attempted by the couchbase callback which does not validate
the certificate name.
Also NCBC-3992 (see that ticket).

Modification
============
If the built-in certifcate validation indicates that the name
validation fails (and the option to ignore is not true), then
the validation fails.
This change also fixes the issue described in NCBC-3992.

Change-Id: If8afedb6714ef9f878aa4c263a4ec5407e9e8205
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/225664
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Emilien Bevierre <emilien.bevierre@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
(cherry picked from commit af8e1fa08c70a70021363334fa781c9292b39255)
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/227059
Tested-by: Emilien Bevierre <emilien.bevierre@couchbase.com>