NCBC-4002: OOM when streaming in .NET Framework 4.8
Motivation
==========
We hit OOM when streaming large query responses with builds
like net6 or net8, which was recently fixed.  But the Framework
4.8 (and probably lower) still have the issue.

Modification
============
Older .NET (before net6) didn't handle buffers as well as more
recent builds.   So - when we call SendAsync on a HttpClient, we
cannot use the DefaultCompletionOptions.   When doing so, the
older client will try to read the entire response into one buffer,
which of course gives us the OOM when we hit 2Gb.  The solution
seems to be just tell it to use ResponseHeadersRead, at which point
we read the rest as a stream and we are ok.

So - since this will require the user Dispose of the QueryResult
or leak HTTP connection in older .NET, we will add an option to
the QueryOptions which controls whether or not we use the
ResponseHeadersRead option, and that defaults to false if the
build is pre NET 5.0.
Results
=======
On my windows box with Framework 4.8, this fails without this
change (trying to stream 7.5Gb or so), and works fine with the
change.

Change-Id: I55a3e364cbed2b67b3889dc7bf83e850c6795464
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/226343
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>
(cherry picked from commit e969b9c7a0db3ee5ad37a9e9b6e871ccff0118ba)
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/227057
Tested-by: Emilien Bevierre <emilien.bevierre@couchbase.com>
Reviewed-by: Emilien Bevierre <emilien.bevierre@couchbase.com>