NCBC-4012: Suport ServerFeature SubdocBinaryXattr
Motivation
==========
This will allow us to (with changes in transaction code) support
binary documents in transactions.

Modification
============
Added the new ServerFeature to the set of ServerFeatures.   Then,
we need to negotiate it as the others.

We will add a helper in transactions to get this from the
ICouchbaseCluster we are passed, and not add any sort of helper
in the client itself.

Results
=======
Correctly notices when it is/isn't supported by the server.

Change-Id: Ia77174dc83f067da77769e324aa6e090e2bc3db6
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/227121
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>