NCBC-3738 Support Binary Objects in Transactions
Motivation
==========
Get .NET to pairity with other transaction sdk implementations.

Modificatation
==============
This is strictly the non-transactional changes.   We have added the
ability to push binary data to an xattr.   Note for now this in an
internal feature - will perhaps be public later.   But the coming
changes in transactions which use this will be public.

Results
=======
Tests pass.

Change-Id: I462ef1cc10a8985ca4e7b79c12beb0e319cad555
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/228257
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>