NCBC-3738: extBinarySupport for transactions
Motivation
==========
Add ability to have transactions that involve binary docs.

Modification
============
We now have an options class and builder for GetAsync, InsertAsync,
and ReplaceAsync operations.   They all take an optional
transcoder and parent span.   Under the hood, we use the optional
transcoder to properly encode the transactional metadata, so that
we can transact across binary docs.   This relies on the server
support for binary xattrs.

Results
=======
All (but one) of the extBinarySupport tests pass.  There will be
work on transcoders in transactions coming shortly to address that
one failure (and some other issues not currently tested for).

Change-Id: Ia37c7e2285ab9cc8c60cc7382c8cafb3d41b9217
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/228472
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>